requirements.txt
yfinance>=0.2.28
pandas>=2.0.0
numpy>=1.24.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
slack-sdk>=3.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.100.0
pyyaml>=6.0
pytest>=7.4.0
pytest-cov>=4.1.0
requests>=2.31.0
